refactor: 자판기 탭 클래스 상속
탭 클래스의 중복되는 메서드 및 멤버 변수를 상속을 통해 중복되지 않도록 수정함.
